+++ mehari/src/db/check/mod.rs
@@ -14,7 +14,7 @@ use std::path::{Path, PathBuf};
 use strum::Display;
 
 use crate::annotate::seqvars::load_tx_db;
-use crate::db::create::Reason as FilterReason;
+use crate::db::create::models::Reason as FilterReason;
 use crate::pbs::txs::{TranscriptDb, TranscriptTag};
 
 /// Check a mehari transcript database against information from
